% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IHE_style.R
\docType{data}
\name{IHEstyle}
\alias{IHEstyle}
\alias{IHE_BlueGray}
\title{IHEstyle}
\format{

}
\usage{
Add + IHE_style to your ggplot2 function in order to use the package.
}
\description{
This package allows you to add the IHE theme to your ggplot graphics.
}
\examples{
library(ggplot2)
ggplot2::ggplot(mpg, aes(displ, cty, colour = class)) +
geom_point() +
IHE_style()
}
\author{
Gunnar Brådvik, email = gunnar.bradvik@ihe.se
}
\keyword{IHE_style}
