% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IHE_style.R
\docType{data}
\name{IHEstyle}
\alias{IHEstyle}
\alias{IHE_BlueGray}
\alias{IHE_style}
\title{IHEstyle}
\format{
An object of class \code{character} of length 1.
}
\usage{
This package allows you to add the IHE theme to your ggplot graphics.
}
\description{
This package allows you to add the IHE theme to your ggplot graphics.
}
\examples{
ggplot(mpg, aes(displ, cty, colour = class)) +
geom_point() +
IHE_style()
}
\author{
Gunnar Brådvik, email = gunnar.bradvik@ihe.se
}
\keyword{IHE_style}
