% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.r
\name{IHEstyle}
\alias{IHEstyle}
\title{IHEstyle}
\usage{
Add + IHEstyle2023 to your ggplot2 function in order to use the package.
}
\description{
This package allows you to add the IHE theme to your ggplot graphics. Note that we have saved the old styles in the package for backwards comparability. The current style is IHEStyle2023.
}
\examples{
library(ggplot2)
library(ggExtra)
library(ggthemes)
library(tidyverse)
library(extrafont)
ggplot2::ggplot(mpg, aes(displ, cty, colour = class)) +
geom_point(size = 3) +
IHEstyle2023()
}
\author{
Gunnar Brådvik, email = gunnar.bradvik@ihe.se
}
\keyword{IHEstyle2023}
